refactor: In절 제거 및 id로 각각 조회, api 명세에 맞춰 RequestMapping 수정
