+++ src/main/java/nextstep/subway/dto/LineRequest.java
@@ -1,7 +1,5 @@
 package nextstep.subway.dto;
 
-import java.util.List;
-import javax.persistence.EntityNotFoundException;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Station;
 
@@ -58,18 +56,15 @@ public class LineRequest {
         return upStation;
     }
 
+    public void setUpStation(Station upStation) {
+        this.upStation = upStation;
+    }
+
     public Station getDownStation() {
         return downStation;
     }
 
-    public void isUpStationThenSet(List<Station> stations) {
-        this.upStation = stations.stream().filter(station -> station.getId().equals(upStationId))
-                .findFirst().orElseThrow(EntityNotFoundException::new);
-    }
-
-    public void isDownStationThenSet(List<Station> stations) {
-        this.downStation = stations.stream()
-                .filter(station -> station.getId().equals(downStationId)).findFirst()
-                .orElseThrow(EntityNotFoundException::new);
+    public void setDownStation(Station downStation) {
+        this.downStation = downStation;
     }
 }
