+++ src/main/java/nextstep/subway/ui/LineController.java
@@ -4,8 +4,11 @@ import java.net.URI;
 import java.util.List;
 import javax.persistence.PersistenceException;
 import nextstep.subway.application.LineService;
+import nextstep.subway.application.SectionService;
 import nextstep.subway.dto.LineRequest;
 import nextstep.subway.dto.LineResponse;
+import nextstep.subway.dto.SectionRequest;
+import nextstep.subway.dto.SectionResponse;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -21,11 +24,28 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/lines")
 public class LineController {
-
     private final LineService lineService;
 
-    public LineController(LineService lineService) {
+    private final SectionService sectionService;
+
+    public LineController(LineService lineService, SectionService sectionService) {
         this.lineService = lineService;
+        this.sectionService = sectionService;
+    }
+
+    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<LineResponse>> showLines() {
+        return ResponseEntity.ok().body(lineService.findAllLines());
+    }
+
+    @GetMapping("/{lineId}/section")
+    public ResponseEntity<List<SectionResponse>> showSections(@PathVariable("lineId") Long lineId) {
+        return ResponseEntity.ok().body(sectionService.findResponsesByLineId(lineId));
+    }
+
+    @GetMapping(value = "/{name}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<LineResponse> showLineByName(@PathVariable("name") String name) {
+        return ResponseEntity.ok().body(lineService.findResponseByName(name));
     }
 
     @PostMapping
@@ -36,14 +56,11 @@ public class LineController {
         return ResponseEntity.created(URI.create("/lines/" + id)).body(lineResponse);
     }
 
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<LineResponse>> showLines() {
-        return ResponseEntity.ok().body(lineService.findAllLines());
-    }
-
-    @GetMapping(value = "/{name}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LineResponse> showLineByName(@PathVariable("name") String name) {
-        return ResponseEntity.ok().body(lineService.findResponseByName(name));
+    @PostMapping("/{lineId}/section")
+    public ResponseEntity addSection(@PathVariable("lineId") Long lineId,
+                                     @RequestBody SectionRequest sectionRequest) {
+        sectionService.addSection(lineId, sectionRequest);
+        return ResponseEntity.ok().build();
     }
 
     @PutMapping(value = "/{name}")
@@ -58,7 +75,7 @@ public class LineController {
         return ResponseEntity.noContent().build();
     }
 
-    @ExceptionHandler(PersistenceException.class)
+    @ExceptionHandler({PersistenceException.class, IllegalArgumentException.class})
     public ResponseEntity handleIllegalArgsException() {
         return ResponseEntity.badRequest().build();
     }
